+++ index.js
@@ -3,3 +3,4 @@ exports.Actor = require('./src/Actor');
 exports.Stream = require('./src/Stream');
 exports.Step = require('./src/Step');
 exports.Loop = require('./src/Loop');
+exports.Cascade = require('./src/Cascade');