+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+## 1.3.0
+- Added Cascade class
+- Stream supports batch
+- abortListeners now an array like the other listeners
+- startListeners called AFTER first step is called
+
 ## 1.2.0
 - Bug fix action.pause()
-- Added "async" option for Loop class
 - Node22